requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.6
certifi==2018.4.16
cffi==1.11.5
Click==7.0
cryptography==2.4.2
Flask==1.0.2
future==0.16.0
humanize==0.5.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysql-connector-python==8.0.6
netmiko==2.3.0
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
pymongo==3.7.2
PyMySQL==0.8.0
PyNaCl==1.3.0
pyserial==3.4
python-dateutil==2.7.3
python-dotenv==0.8.2
python-telegram-bot==10.1.0
pytz==2018.4
PyYAML==3.13
redis==3.1.0
scp==0.13.0
six==1.11.0
textfsm==0.4.1
virtualenv==16.0.0
Werkzeug==0.14.1
